refactor: files optimised refactored code to respected files
