+++ src/pages/admin/AdminUsers.tsx
@@ -3,7 +3,7 @@ import UserTable from "@/components/admin/UserTable"
 import Pagination from "@/components/common/Pagination"
 import { Input } from "@/components/ui/input"
 import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group"
-import { useGetAllUsers } from "@/hooks/admin/useGetAllUsers"
+import { useGetAllUsers } from "@/hooks/user/useUser"
 import { useEffect, useState } from "react"
 
 const AdminUsers = () => {
